Refactored .js to .jsx
Refactored react componentes, pages and screens to .jsx
